Crate ei_sys

Low level bindings for ei, a library to handle the Erlang external term format and to communicate with distributed Erlang nodes.

Structs

ErlConnect

ei_cnode

ei_x_buff

erlang_char_encoding

erlang_fun

erlang_msg

erlang_pid

erlang_port

erlang_ref

erlang_trace

Constants

EI_MAXALIVELEN

EI_MAXHOSTNAMELEN

EI_MAX_COOKIE_SIZE

ERLANG_ASCII

ERLANG_LATIN1

ERLANG_UTF8

ERL_CONNECT_FAIL

ERL_ERROR

ERL_MSG

ERL_NO_DAEMON

ERL_NO_PORT

ERL_NO_REMOTE

ERL_TICK

ERL_TIMEOUT

MAXATOMLEN

MAXATOMLEN_UTF8

MAXNODELEN

Statics

ei_x_extra
The initial capacity of an ei_x_buff in bytes, when created with [ei_x_new].

Functions

__erl_errno_place⚠
Returns a pointer to the thread-local storage where erl_errno is stored.

ei_accept⚠

ei_accept_tmo⚠

ei_connect⚠

ei_connect_init⚠

ei_connect_tmo⚠

ei_connect_xinit⚠

ei_decode_atom⚠

ei_decode_atom_as⚠

ei_decode_binary⚠

ei_decode_boolean⚠

ei_decode_char⚠

ei_decode_double⚠

ei_decode_fun⚠

ei_decode_list_header⚠

ei_decode_long⚠

ei_decode_longlong⚠

ei_decode_map_header⚠

ei_decode_pid⚠

ei_decode_port⚠

ei_decode_ref⚠

ei_decode_string⚠

ei_decode_term⚠

ei_decode_trace⚠

ei_decode_tuple_header⚠

ei_decode_ulong⚠

ei_decode_ulonglong⚠

ei_decode_version⚠

ei_encode_atom⚠

ei_encode_atom_as⚠

ei_encode_atom_len⚠

ei_encode_atom_len_as⚠

ei_encode_binary⚠

ei_encode_boolean⚠

ei_encode_char⚠

ei_encode_double⚠

ei_encode_fun⚠

ei_encode_list_header⚠

ei_encode_long⚠

ei_encode_longlong⚠

ei_encode_map_header⚠

ei_encode_pid⚠

ei_encode_port⚠

ei_encode_ref⚠

ei_encode_string⚠

ei_encode_string_len⚠

ei_encode_trace⚠

ei_encode_tuple_header⚠

ei_encode_ulong⚠

ei_encode_ulonglong⚠

ei_encode_version⚠

ei_get_tracelevel⚠

ei_get_type⚠

ei_print_term⚠

ei_publish⚠

ei_publish_tmo⚠

ei_receive⚠

ei_receive_encoded⚠

ei_receive_encoded_tmo⚠

ei_receive_msg⚠

ei_receive_msg_tmo⚠

ei_receive_tmo⚠

ei_reg_send⚠

ei_reg_send_tmo⚠

ei_rpc⚠

ei_rpc_from⚠

ei_rpc_to⚠

ei_s_print_term⚠

ei_self⚠

ei_send⚠

ei_send_encoded⚠

ei_send_encoded_tmo⚠

ei_send_reg_encoded⚠

ei_send_reg_encoded_tmo⚠

ei_send_tmo⚠

ei_set_compat_rel⚠

ei_set_tracelevel⚠

ei_skip_term⚠

ei_thisalivename⚠

ei_thishostname⚠

ei_thisnodename⚠

ei_x_append⚠

ei_x_append_buf⚠

ei_x_encode_atom⚠

ei_x_encode_atom_as⚠

ei_x_encode_atom_len⚠

ei_x_encode_atom_len_as⚠

ei_x_encode_binary⚠

ei_x_encode_boolean⚠

ei_x_encode_char⚠

ei_x_encode_double⚠

ei_x_encode_empty_list⚠

ei_x_encode_fun⚠

ei_x_encode_list_header⚠

ei_x_encode_long⚠

ei_x_encode_longlong⚠

ei_x_encode_map_header⚠

ei_x_encode_pid⚠

ei_x_encode_port⚠

ei_x_encode_ref⚠

ei_x_encode_string⚠

ei_x_encode_string_len⚠

ei_x_encode_trace⚠

ei_x_encode_tuple_header⚠

ei_x_encode_ulong⚠

ei_x_encode_ulonglong⚠

ei_x_encode_version⚠

ei_x_format⚠

ei_x_format_wo_ver⚠

ei_x_free⚠

ei_x_new⚠
Initializes a new ei_x_buff by filling all fields and preallocating a buffer of size ei_x_extra.

ei_x_new_with_version⚠

ei_xconnect⚠
Connects to an Erlang node, without a timeout.

ei_xconnect_tmo⚠
Connects to an Erlang node.

ei_xreceive_msg⚠
Receives an Erlang message from the connection setup on the file descriptor fd.

ei_xreceive_msg_tmo⚠

free_fun⚠